requirements.txt
amqp==5.1.1
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
boto3==1.26.81
botocore==1.29.81
Brotli==1.0.9
CacheControl==0.12.11
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==39.0.0
cssselect==1.2.0
cssselect2==0.6.0
daphne==4.0.0
Deprecated==1.2.13
distlib==0.3.5
Django==4.0.6
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-filter==22.1
django-mathfilters==1.0.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
filelock==3.7.1
fonttools==4.37.1
future==0.18.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
joblib==1.2.0
kombu==5.2.4
lxml==4.9.1
Markdown==3.4.1
MouseInfo==0.1.3
msgpack==1.0.4
nexmo==2.5.2
numpy==1.23.1
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
parsel==1.7.0
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.36
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
PyDispatcher==2.0.7
pydyf==0.2.0
PyGetWindow==0.0.9
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pymongo==3.12.3
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
PyPDF3==1.0.6
pyperclip==1.8.2
pyphen==0.13.0
PyRect==0.2.0
PyScreeze==0.1.28
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
pytweening==1.0.4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywebcopy==7.0.2
PyYAML==6.0
qrcode==7.3.1
queuelib==1.6.2
razorpay==1.3.0
reportlab==3.6.11
requests==2.28.1
requests-file==1.5.1
rfc3986==1.5.0
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
Scrapy==2.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlparse==0.2.4
svglib==1.4.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tldextract==3.4.0
tqdm==4.64.1
twilio==7.13.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
vine==5.0.0
virtualenv==20.15.1
virtualenvwrapper-win==1.2.7
w3lib==2.1.1
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.8
xlwt==1.3.0
zope.interface==5.5.2
zopfli==0.2.1
